Add remove-at-point. Remove restriction for maintain when undo-in-progress
The original text-clone--maintain explicitly restricted sync when undo is in progress
It is in the `and' clause, and I removed that part of condition. From my test, the restult
is working fine. There is no documentation I can see as to why undo-in-progress was
specifically restricted. It is more intuitive if undo can also sync.