Re-architect for real-time sync using custom text-clone functions (#4)
* Add text-clone logic adjusted to work with org-transclusion

* Fix text-clone--maintain where it does not work when the source is narrowed

* Fix bi-directional sync -- now it works both ways

* Remove update logic from update-src-at-point. Now real-time sync does it

* (UX) Fix update-at-point to not to move point

* (WIP) Start refactoring

* (WIP) Still refactoring to have fn list to work

* (WIP) parameterization seems to be getting somewhere

* Fix bi-directional real time sync (WIP, but works)

* Add remove-at-point. Remove restriction for maintain when undo-in-progress

The original text-clone--maintain explicitly restricted sync when undo is in progress
It is in the `and' clause, and I removed that part of condition. From my test, the restult
is working fine. There is no documentation I can see as to why undo-in-progress was
specifically restricted. It is more intuitive if undo can also sync.

* Remove fn yank-src-to-target. It's obsolete

* Replace update-src-at-point with save-src-at-point

As now the clones are syncronised real-time, it is no longer necessary to update
the source buffer. Saving it is still necessary, along with the backup mechanism.

* Replace update-src-at-point with save-src-at-point

As now the clones are syncronised real-time, it is no longer necessary to update
the source buffer. Saving it is still necessary, along with the backup mechanism.

* Refactor done

* Fix defect where an additional line is inserted every time add operation is done

Cause: the transclusion link is deleted, but the line it used to occupy remains
This line needs to be removed.

Solution: delete-char is added to do so when transclusion link is deleted.