Add support for paragraph with using dedicated link
Supports this form otc:file:path/to/file.org::headline

The target paragraph must be identifiable by a dedicated link with
a <<paragraph-id>>: e.g.

   Lorem ipsum dolor sit amet, consectetur adipiscing elit.
   Suspendisse ac velit fermentum, sodales nunc in,
   tincidunt quam. <<paragraph-id>>

It is generally assumed that the paragraph-id is placed after its content,
but it is not an absolute requirement; it cna be in the beginning (before
the content) or in the middle of it. It uses `org-link-search' to locate
the <<dedicated-link>> and use `mark-paragraph' to select the content.