Remove transclusion ov non-sticky at rear
Text-property rear-nonstky does not work as intended.
It works per stentence (or some unit of text), not the overlay. So you end up being
able to change the change after the sentence.

What I need to achieve is the ability to make the region right after the overlay editable.
Currently that is not the case.