Add/fix transcluding content for ID
When ID is used as a link, a special treatment seems to be necessary.
When the source buffer is narrowed and the relevant part is invisible, ID does not move the point
to the correct part where ID is pointing to. It is necessary to call org-id-goto again.

This is not the case for custom ID -- it is part of the file link and search-option (::custom-id).
No problem identified for cutom ID (tested).