Add control for what elements to exclude
This currently works only for transluding an org mode buffer via a file link
wihtout search option.

The next step is to make the list of excluded elements to be a customizing
variable. It shoudl be a list.

The only exception from this is the "first section", which is a section before
the first headline. It includes the buffer/file properties defined with
keywords such as `#+title' and `#+author', etc.

Exclusion of the first section would need to be another customizing variable,
probably a boolean flag.