Fix org-roam-id-open (WIP)
The immediate issue was that org-roam-id-open "overrides" the
org-open-at-point.  This is done via Org Mode's hook
`org-open-at-point-functions` and should be expected.

This is a good opportunity to refactor the way `org-transclusion-open-link-*` works.
Now I can use `org-open-at-point-functions` hook, and can remove the advice-add.

`advice-add` is not a good solution as it is a global change, where
Org-transclusion is buffer local.  With the hook, it can be a buffer local.

This commit is a WIP; need to clean up the superflous functions I added before realising hook is a better solution than advice.