Fix metaup/down
In trying to fix the issue of multiple headlines within a transclusion, I
inadvertently added a problem. The logic scaned all the buffer after the
transclusion in question, thus changing all the headlines regardless of it
being in the same transclusion or not -- non-transclusion ones too! Not good.

The issue of multiple headlines within one transclusion still remains.