fix: table misalignment.
This is meant to fix GitHub issue #54.

As far as I could see, when a table contains a
link, the table does not seem to be aligned
automatically when transcluded. As the transcluded
region becomes read-only, it cannot be re-aligned
after transclusion is done.

This solution is not great, as it only deals with
a single table transcluded; i.e. single
element. When a whole buffer is to be transcluded
with multiple tables, this solution will not work.

As a generic solution including both a buffer and
a single elemnet, I wanted to align tables before
insertion. For some reason, this does not seem to
work with the `with-temp-buffer` solution. I don't
have enough time to investigate what causes this.