fix: Re-align multiple tables in a buffer
This is a generic fix for #54; it works both for a
single element (e.g. table, or dynamic block with
a table, etc.) and for multiple tables included
in a buffer.

The issue was that the temp buffer used for
formating was not in the org-mode, thus the entire
link `[[link][description]]` was expanded before
alignment, which leads to misalignment when
content ended up in an org buffer.

It has been tested that this new way works for
both states of org-toggle-link-display.

changes. Lines starting