add: buffer-substring-advice-enabled
Decided on taking an approach described in Emacs in-system manual (cl)Obsolete
Features > Obsolete Macros.

The issue is that I think it's necessary to override
`buffer-substring-no-properties` used in `org-element` parsers. It is a widely
used function, so limiting the scope has been critical.

The approach is to use `advice` with a dynamically scoped variable to limit
the case when advised `buffer-substring-no-properties` is to be used.

The scope of the variable is global (dynamically scoped) but it's a local
variable, so the custom logic kicks in only for the buffer in question (tested
with scratch buffer while stopping at a break point within the custom logic in
question).

Lexically scoped function replacement with `cl-flet` doest not work, as the
function is used further down in the call stack in `org-element` parser
functions.

Dynamically scoped `cl-letf` to replace the function definition with
`(symbol-function 'fn)` works, but it is essentially equivalent to
`advice-add` -- both globally applicable (dynamically scoped).

As Emacs is single-threaded, the risk of other functions inadvertently calling
the custom logic should be low, but I think it's prudent to minimize this
risk.