fix: content-insert with temp buffer, fix #68
Fixes #68 and more subtle rendering issue of transcluded content.

I had refactored the way to insert transcluded content in the buffer. I newly
introduced temp-buffer, which caused the issue of the content to be rendered
incorrectly as an Org subtree (this was only a problem of subtree).

Fix was to turn on Org Mode (with delayed mode hooks) in the temp buffer.