chg: link syntax & create-from-link features
The link syntax #+transclude was unable to interpret the link when double
quotation marks are included in the search option. For example, if the
headline title was `* The "denotating phrases"` (example below) with these
double quotation marks, add-at-point fails because it was not intepreted as a
valid link.

    #+transclude: t "file:test-content-first-section.org::*The "denoting phrase""

Solution:
Change the syntax to this:
: #+transclude: t [[file:test-content-first-section.org::*"On Denoting"]]

Advantages
- This takes care of the quote issue
- in the keyword, =C-c C-o= (standard "follow link" =org-open-at-point=) works
- Easy to work with existing links in Org files

Disadvantages:
- It breaks away from the =#+include= syntas of using double quotations

I think the advantages outweight the disadvantages I can think
of. Transclusion is similar but not extention to the include syntax.

In addtion, relatedly, `create-from-link` now has a new feature.  You can pass
a digit argument 1-9 to specify the :level property.