fix: live-sync get-enclosing-element
The issue was that the `get-enclosing-element` was incorrectly identified
region of a paragraph when a transcluded paragraph was placed right next to
another paragraph present in the buffer. As there was no blank line
in-between, the transcluded paragraph was subsumed by the surrounding one.

In order to void this problem, this function nows works in a temporary org
buffer to isolate the transcluded region and source region from the rest of
the original buffer.

TODO: For source buffer, this function works on overlay. It feels a bit
fragile, and inconsistent with the way translused region works.