add-feat: live-sync buffer management
Add a couple of functions to manage the live-sync edit buffers.

Key principle is that there must be only one live-sync region active globally
in an Emacs session to support deliberate editing experience.

In line with this principle, live-sync also re-arranges the windows to show
only two windows, removing the others. The buffer that does the transcluding
and editing is displayed with the source buffer being edited.

Upon existing live-sync, Org-transclusion attempts to recover the window
configuration as it was before live-sync.