add: org-transclusion-search-or-add-next-empty-line
Fixing an edge case issue where there is no empty space at the bottom line of
a buffer, `create-from-link` does not add a new line but add the keyword at
the end of the line. As the keyword can only be in an empty line (with or
without inital margin, space, tab, etc.), transclusion does not work.