add: insert-behind-hooks and insert-before-hooks
Behind and front hooks seem necessary to deal with the case when the whole
regin is deleted and undone, or similar.

In order to deal with the case of deleting a region that includes the
beginning of the overlay, it is also helpful to make overlay with
FRONT-ADVANCE nil so that characters inserted at the beginning of the ovelay
is to be included in the overlay.

When live-sync starts, the read-only text-prop is to be removed from beg - 1
in line with the other changes above.