fix: possible to place transclude w/o a blank line
Issue was when transclusion keywords are placed without a blank line,
remove-at-point deletes one of the transclusion keyword. This is because when
you call remove-at-point sequentialy, the marker would delete the transcluded
text and include the keyword; this keyword would get included in the second
transcluded region via the marker and subsequently deleted along with the
content of the second transclusion. Only the keyword of the second
transclusion would survive.

This is now fixed by using `insert-before-markers` instead of `insert`.