add: remove live-sync overlays when deleted
Adding a mechanism to remove both of the live-sync overlays (transclusion and
source) when transclusion is completedly deleted. This solves the problem of a
source overlay to be orphaned in such cases.

TODO: after overlays have been removed properly, you can undo transclusion
buffer to recover the text deleted. This is fine, but the text region is no
longer read-only