fix: kill-buffer/emacs
Before this fix, when you use live-sync and then either kill-buffer or
kill-emacs, transclusion text regions remain as the body of the text; not
returning to keywords.

This was because live-sync temporarily turns off before- and after-save-buffer
hooks to keep the edititng mode while you edit. This resulted in the buffer saved in filesystem to retain the text even when live-sync was exited.

This commit fixes the issue and clear files in the filesystem of the
transcluded text content.