chg: save-buffer hooks
Introducing local variable `org-transclusion-before-save-transclusions`

This variable is used to remember the active transclusions before
`save-buffer`.  It is meant to be used to keep the file the current buffer is
visiting clear of the transcluded text content.  Instead of blindly deactivate
and activate all transclusions with t flag, this variable is meant to provide
mechanism to deactivate/activate only the transclusions currently in effect to copy
a text content.