feat: src-lines
This logic from include seems to work.  I will consider if "excluded" makes
sense -- 5-10 probably should include line 10, which to me is more intuitive.

```
[From (org) Include files]

   Inclusion of only portions of files are specified using ranges
parameter with ‘:lines’ keyword.  The line at the upper end of the range
will not be included.  The start and/or the end of the range may be
omitted to use the obvious defaults.

‘#+INCLUDE: "~/.emacs" :lines "5-10"’   Include lines 5 to 10, 10 excluded
‘#+INCLUDE: "~/.emacs" :lines "-10"’    Include lines 1 to 10, 10 excluded
‘#+INCLUDE: "~/.emacs" :lines "10-"’    Include lines from 10 to EOF
```