fix: remove-at-point when two trs are adjacent
Use of insert-before-markers solves the problem of
removing adjacent transclusions without blank
lines for one direction -- top to bottom. The
problem remained without being noticed for the
other order, bottom to top.

The surface problem was that removing two
transclusions from bottom to top deletes one of
them. This was only a problem when these
transclusions were placed immediately after
another with no blank line in-between.

The cause of the problem was the markers stored as
a text property, tc-end-mkr of the transclusion
immediately before the one being removed. The
diagram illustrates this. Removing tc2 first means
that the #+transclusion keyword gets put back. Now
because of the marker's type, the marker moves to
include the keyword. Thus removing tc1 then
removes tc1 text *and* the keyword for tc2.

---- BUFFER beg ----

-- tc-beg-mkr (tc1)
    tc1 text
-- tc-end-mkr (tc1) & tc-beg-mkr(tc2) on the same point
    tc2 text
-- tc-end-mkr (tc2)

---- BUFFER end ----

Chaning the markers type alone didn't solve this
issue. This is mostly likely because we use
`insert-before-marker`. `insert` could be used
instead, but that will cause the other order the
same issue. One of the direction seems to have an
issue either way.