Initial commit for more-search-option branch
based on the current main branch, broke live-sync?

add: merge org-transclusion-link-open-org-file-links with
org-transclusion-link-open-other-file-links to one
org-transclusion-open-file-link that always call
org-transclusion-content-get-from-file-link to build an appropriate
argument for the file link to pass onto
org-transclusion-content-get-buffer-or-element, what it does for every
file link is, operate on search option by pointing to the appropriate
section of the source buffer determined by its argument, if it's an
org file, call
org-transclusion-content-get-org-buffer-or-element-at-point. For
non-org file, just return the information about the section it's
pointing to.

add: two customizable variables org-transclusion-line-search-style &
org-transclusion-regexp-search-function both are used for dealing
with their representative search option.