add :disable-auto; chg :only-contents
`:disable-auto` ::
No need for a value (e.g. t, nil, etc.). When this property is in the keyword,
`org-transclusion-add-all-in-buffer` will not transclude it. No effect on
manual transclusion with add-at-point

`:only-contents` ::
`:only-content` (singular) is also accepted. No need for a value. Ignores the
headlie's title, and transclude only content of the headline. When it contains
a subtree (sub-headline): the hirarchy is ignored and include the contents of
all the subtree as well.