content-format-org: fix formating of text files
Add function org-transclusion-content-format-org
Add function org-transclusion-type-is-org
Adjust function org-transclusion-content-format

The issue before this change was that formating of the text file (including
source files) didn't consider the indententation.  Most notably, this resulted
in the live-sync not correctly working.

The cause of this issue was the default content-format function only considered
org files and turned on org-mode within it -- thus removing intial indentation
(depending on the user's configuration).

The new default content-format function now does not turn on org-mode and
returns the content as is.  The new format function specific to org buffers is
introduced; it's logic is identical to the previous default.

The new type-is-org function is used to determine the org-transclusion-type.