fix Check lenghs of overlays before live-sync
The overlays for live-sync must be identical; otherwise, live-sync changes the
source buffer in an unpredictable way and often removing too many characters.

Now check of the lengths is implemenented. This aborts the process with
user-error. The sequence of the statemetns around this check is rearranged in
order to avoid inconsistency with before/after-save-hooks and other variables.

Essentially, this check needs to be done first before hooks and variables are
changed.