feat: :set prop for extentions
org-transclusion-set-extensions function added to require features defined as an
extention. org-transclusion-indent-mode is a new extention using this feature.

Adding an exntesion will require the feature.

Removing one, however, does not "un-"require the feature, so it may not turn off
the effect -- as can be seen with indent-mode, it needs to be explicitly coded
to avoid the effect.

Restarting Emacs will be required to turn off extentions.