fix: #93 open-source-marker
`org-translusion-open-source-marker` was using
`org-transclusion-live-sync-enclosing-element` to get the org-element at
point. This led to an error message "Live sync cannot start here", which is not
relevant for the open-source function.

This function is now replaced with `org-element-context` which should be equivalent in the context of the open-source function.

My Regression tests confirm this, too.