fix: transclusion within another transclusion #91
transclusion inside another. This should not happen by design.

This commit add `org-transclusion-check-add` and in it a new check has been
added to prevent this case. Currently it checks the following two cases:

Case 1. Element at point is NOT #+transclude:
Case 2. #+transclude inside another transclusion"

The PR #96 attempts go beyond this design. It currently seems to complicate
the cases for `org-transclusion-remove`. We might revist the case for
nesting (a special case of transclusion within another transclusion);
however, I am not prepared to go through regression and unaticipated issues
with this major change of assumptions.