fix #97: edge case, capture buf transcludes
first. To deal with an edge case where transclusions were added to a capture
buffer -- e.g. `org-capture' or `org-roam-catpure' --, check for
`buffer-file-name' to see if there is a file visited by the buffer. If a "temp"
buffer, there is no file being visited. In this case, don't attempt save the
buffer.