doc:WIP:testing doc generation of ELPA
When org-transclusion.texi is missing, README.org is exported to *.texi
file. ELPA then generates *.info and *.html from it. You can define "*" in
README.org with `#+export_file_name:` meta data. It is not clear what happens
when a filename `org-transclusion.texi` already present in the repository?

In this commit, org-transclusion.org and org-transclusion.texi is present with
license preambles. README.org is specified as the doc by ELPA setting.

Which file is used for the *.html and *.info?