add:inhibit-read-only: fix:Org export errors with read-only
Some Org exports backends (e.g. texinfo) use org-babel, which does not work when
part of the buffer is read-only andoutputs an error message.

Add `org-transclusion-inhibit-read-only` function and add it locally to
`org-export-before-processing-hook`. It temporarily inhibits read-only for the
buffer for export.

With this, `manual-gen.el` for make is no longer needed, replaced with an
Elisp expression instead.