fix:#105 Remove use of silent-mod macro & inhibit-modification-hook
It seems to be that org-element--cache has trouble synching cache when
inhibit-modification-hook is t. Removed.

There is one place where with-silent-modificaiton is used but it's a proper way
where only text props are modified with no change to the text content in the
buffer, thus no change to the Org element's tree. This does not seem to affect
the cache sync.