fix:Another Emacs remove-all leaves superfluous text prop after kill
Fix a very rare but destructive bug.

Repro:
1. Launch Emacs session
2. Visit an Org file with transclusions
3. `org-transcludion-add-all`
4. Keep it open. Save to remove `buffer-modified-p` to return nil

5. Launch a second Emacs session
6. Visit the same Org file with transclusions
7. `org-transcludion-add-all`
8. While the transclusions are still on, kill the buffer or kill Emacs

10. Back in the first Emacs session, observe that all the transclusions of the
first session are now removed

11. `org-transclusion-remove-all` again

12. A large chunk of text disappears, especially when the last transclusion
occupies a large region

The last transclusion leaves the text properties only one point at end of the
`#+transclude` line, keeping all the text properties from when the transclusion
was on. With this one-point text properties, remove-all removes all the region
that the transclusion used to be, removing a large chunk of non-transclusion text
content.

This was a problem especially when eshell session was used to generate
documentation while it is being edited.

Rare bug, probably only for the author of the package. High impact, nonetheless.