fix: follow-up on the previous fix
The previous one does not address the case where the second Emacs session was a
manual one not batch job for make, etc.  The second call to remove-all is good
as it address the issue of live-sync mentioned in the docstring as well as this
new case that is now identified for the second Emacs session.