fix:add:line-beginning-position
Within add, the beginning of position for the begining of transcluded content
inserted is determined by where the point is at the time of calling the add
function.  This can be anywhere in the \#+transclude line.  This should be
standardized to the beginning of the line. Otherwise, the text-prop may not be
added to the text-props for indent-mode, or the correct begin may not be passed
to the after-add-hook, which is to be added here.