feat: org-transclusion-after-add-hook
Refer to discussions in PR #119.

A new abnormal hook, `org-transclusion-after-add-functions` has been added.

"Functions to be called after a transclusion content has been added.  The hook
runs after the content and the read-only text property have been added so it is
not supposed to manipulate the content but to add further text properties.  For
example, it is used by the `org-transclusion-indent-mode' extension to support
`org-indent-mode'.  The functions are called with arguments beg and end,
pointing to the beginning and end of the transcluded content."

The existing extension, `org-transclusion-indent-mode` has been refactored to
use the newly added hook.  In this refactoring, the part which existed in the
remove function has been removed and tested OK to do so.

Documentation is WIP and to be added before a new version makes its way to ELPA.