fix: conditions in org-transclusion-add-src-lines
The conditions commented during the development were put back in.

Treating an org-id target as a normal text buffer means that even when
the ID points to an Org headline, the ending of the transclusion will be
dictated by the `:lines`  or `:end` properties, not the end of
headline. This means that if the `:lines` property is set to ":line 0-",
the resultant transclusion will be from the initial line of the headline
to the end of the buffer, going beyond the boundary of the headline.