fix: Delete superfluous trailing space on remove
Before the fix, function 'org-transclusion-keyword-plist-to-string'
would add a superfluous trailing space when converting #+transclude
keyword's properties back to the string.  This does not cause any harm
in normal circumstances.

The issue would occur with a rare combination of the case where:

- You have a buffer with a large number of transclusions activated,
  around 50 or more

- You set a 'whitespace-cleanup' in 'before-save' hook

- You have 'auto-save-visited-mode' or similar that automatically saves
  a buffer

If you had combination, and then saved the buffer with the
transclusions, you might get an error mid-way of the save operation. You
would not lose any data in the buffer, but the buffer would not
re-activate all the transclusions that had been active before
buffer-save.

This was because of the superfluous trailing spaces and automatic
removal of them, which caused the mismatch of the point of each
transclusion that Org-transclusion remembered during the save operation.