fix(make-from-link): Change the treatment of universal-argument
Before, the command make-from-link automatically would add the
transclusion when org-transclusion-mode minor mode is active. Passing
universal-argument only would the behaviour when the mode is inactive to
automatically add the transclusion from the newly-added transclusion
keyword.

This fix changes the behaviour as follows and makes `universal-argument`
more useful.

The new behaviour is:

If `org-transclusion-mode' is active in the current buffer, the newly
created transclusion automatically gets added. If the mode is inactive,
the keyword gets simply inserted to the current buffer.

If you pass a `universal-argument', this function reverses this: if the
mode is active, the keyword gets inserted; if the mode is inactive, the
transclusion gets added.

This change is motivated by this discussion on Org-roam discourse:
https://org-roam.discourse.group/t/how-to-do-transclusion-to-org-files-as-fast-as-possible/3002/5?u=nobiot