refactor: content-format-fns and src-lines
The basic intent is to refactor -src-lines.el to separate the scope of
src and lines more clearly. The hope is also to make thing-at-point
separate from both.

To do this, the first thing done in this commit is to separate out the
function to wrap the content into src. This is now done by the abnormal
function hook -content-format-functions.

In order to do this, the hook's function call API has been changed. The
last argument of the unction call in the hook used to be
:current-indentation. This is anyway part of the keyword-values plist,
so the function API now expects the plist instead of just
:current-indentation.