docs: mention "transparent" default face spec
The default for faces 'org-transclusion' and 'org-transclusion-source'
is no color or any other specification (transparent). An intended use of
these faces is when you use packages such as ~org-indent-mode~ or
~org-modern~, which might override the fringe ~org-transclusion~ places.
To make the transclusion and its source visible, you might like to put
some background color to source and transclusion.