fix: Don't ever prompt to create a headline when transcluding
The default setting of `org-link-search-must-match-exact-headline'
causes Emacs to prompt the user to create a new heading when a link
target is not found.

This commit temporarily overrides the default setting to signal an
error when no headline exactly matches the link target.  Since errors
are demoted, this skips the failed transclusion.

However, if the user has set
`org-link-search-must-match-exact-headline' to perform a fuzzy search
for link targets, we don't override the option.