fix: Patch infinite loop by adding a secondary method #177
Collective analysis efforts have found that infinite loop occurs when variables
`beg` and `end` have an identical value in `org-transclusion-remove`.

It's at the top of the function and looks like this:

(beg (marker-position (get-char-property (point)
                                         'org-transclusion-beg-mkr)))
(end (marker-position (get-char-property (point)
                                         'org-transclusion-end-mkr)))

These values are used to know the size of the transclusion to remove it. The
size of overlay on the source cannot be used because filter can alter the text
size. `text-property-search-forward` can be a reliable way to do this.

The message "(org-transclusion) Something is off" is still a WIP version. I
think we need to come up with a better message if the method above is deemed
viable as a preventive measure.