chg: Make hooks to set up extensions removable #261
Added the  following utility functions to support the move

- `org-transclusion-extension-functions-add-or-remove`
- `org-transclusion-extension-set-a-hook-functions`

Modified function `org-transclusion-load-extensions-maybe` so that setting
`org-transclusion-extensions` with the `customize` facilities will automatically
activate the corresponding minor mode for each extension.

Use of `customize` and `org-transclusion-extensions` is optional. Users can
ignore these and use extensions by loading and activating minor modes in their
init.el file.

So, this below should work (each minor mode function has the autoload cookie, so
the library should not have to be explicitly loaded).

```
(with-eval-after-load 'org-transclusion
    (org-transclusion-indent-mode +1))
 ```

Adding to the list and requiring the library like this below did not use to be
necessary...

```
(with-eval-after-load 'org-transclusion
     (add-to-list 'org-transclusion-extensions 'org-transclusion-indent-mode)
     (require 'org-transclusion-indent-mode))
```

The Customizing facility would load the active extension library with
`require`.... But I think this was never intuitive.