fix(add): org-id with search option (::*heading) #237
While fixing the issue, I am extending the "add" operation to allow for *any*
link.  A lot is still under development and many things will change. But this
seems to work without breaking any user-facing commands and function.

Introduced a new function `org-transclusion-add-target-marker`, which helps
transcluding *any* link (as long as Org can reach the link target and Emacs
creates the link's target buffer -- tested with `notmuch` link successfully).

Refactored the code for the "add" operation with this new approach.